server/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
openai
faiss-cpu
streamlit
requests
python-dotenv
